Move hip_api_trace_entry_t to roctracer_trace_entries.h
