+++ inc/roctracer_trace_entries.h
@@ -6,6 +6,7 @@
 
 #include <roctracer_roctx.h>
 #include <roctracer_hsa.h>
+#include <roctracer_hip.h>
 #include <ext/prof_protocol.h>
 #include <ext/hsa_rt_utils.hpp>
 
@@ -49,6 +50,20 @@ struct hsa_activity_trace_entry_t {
   uint32_t pid;
   activity_record_t *record;
   void *arg;
+};
+
+struct hip_api_trace_entry_t {
+  std::atomic<uint32_t> valid;
+  roctracer::entry_type_t type;
+  uint32_t domain;
+  uint32_t cid;
+  timestamp_t begin;
+  timestamp_t end;
+  uint32_t pid;
+  uint32_t tid;
+  hip_api_data_t data;
+  const char* name;
+  void* ptr;
 };
 
 #endif
